server/requirements.txt
djangorestframework-jwt=1.11.0
django~=3.0.3
requests~=2.23.0
djangorestframework~=3.11.0
django-filter~=2.2.0
pillow~=6.2.2
wagtail~=2.8
pymysql~=0.10.1
pytz~=2019.3
redis~=3.5.3